+++ planner-frontend/src/pages/admin/ShiftManagement.css
